PoC: Add PasswordHasher with keyed hashing
Keyed hashing is the solution to prevent a SQLi or database backup from
enabling reversing *any* of your hashes, no matter how simple. Configure
a secret (or set of secrets) on your server, pass this secret when
initializing the PasswordHasher, and an attacker will be unable to
reverse anything without this secret.

Stuff I want to add to make this fully usable:
- Tests covering:
  - Basic use
  - Compatibility with existing PasswordHasher
  - Upgradeable (all parameters should be possible to upgrade)
- Robustness against maliciuous hashes from database (to prevent DDoS)
- Support secret rotation (given a dict of keyid->key pairs, store keyid
  in associated data, automatically load correct key)
- Add paslib-equivalent version of verify_and_update returning a tuple
  (valid, new_hash). This to enable storing a new hash when parameters
  have been bumped.