Use argon2_context() for PasswordHasher
Some of the tests fail, haven't been able to track what's the difference
in these calls yet. The previous PasswordHasher wrapped argon2_hash()
instead of argon2i_hash_encoded and the likes, but should have been the
same. I'm not sure yet what exactly is the difference, but the output
from the refactored PasswordHasher matches that of the argon2 CLI. Will
debug further.